Move schema to entity file
