+++ src/routes/relay/entities/relay.legacy.dto.entity.ts
@@ -1,4 +1,6 @@
-import { RelayLegacyDtoSchema } from '@/routes/relay/entities/schemas/relay.legacy.dto.schema';
+import { AddressSchema } from '@/validation/entities/schemas/address.schema';
+import { HexSchema } from '@/validation/entities/schemas/hex.schema';
+import { NumericStringSchema } from '@/validation/entities/schemas/numeric-string.schema';
 import { z } from 'zod';
 
 export class RelayLegacyDto implements z.infer<typeof RelayLegacyDtoSchema> {
@@ -7,3 +9,10 @@ export class RelayLegacyDto implements z.infer<typeof RelayLegacyDtoSchema> {
   data!: `0x${string}`;
   gasLimit!: string | null;
 }
+
+export const RelayLegacyDtoSchema = z.object({
+  chainId: NumericStringSchema,
+  to: AddressSchema,
+  data: HexSchema,
+  gasLimit: z.string().nullish().default(null),
+});
